api/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
mysqlclient==1.4.6
Pygments==2.3.1
PyMySQL==0.9.3
